requirements.txt
psd-tools>=1.14.0
Pillow>=10.0.0
scikit-image>=0.21.0
aggdraw>=1.3.11

# scikit-image: 增强图层效果合成
# aggdraw: 矢量形状渲染（Shape 图层的正确合成需要此依赖）
